binary to module
